--- src/model/point-model.js
@@ -1,23 +1,0 @@
-export default class PointsModel {
-  #points = [];
-  #destinations = [];
-  #offers = [];
-
-  init(points, destinations, offers) {
-    this.#points = points;
-    this.#destinations = destinations;
-    this.#offers = offers;
-  }
-
-  get points() {
-    return this.#points;
-  }
-
-  get destinations() {
-    return this.#destinations;
-  }
-
-  get offers() {
-    return this.#offers;
-  }
-}
